Move attachment ViewHolders to their own package
