+++ library/src/main/java/com/getstream/sdk/chat/adapter/AttachmentListItemAdapter.java
@@ -6,6 +6,7 @@ import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.getstream.sdk.chat.adapter.viewholder.attachment.BaseAttachmentViewHolder;
 import com.getstream.sdk.chat.view.MessageListView;
 import com.getstream.sdk.chat.view.MessageListViewStyle;
